Removing extensions namespace and moving rect into root folder
