+++ Adventure/Components/OnFire.cs
@@ -2,7 +2,6 @@
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Engine;
-using Engine.Extensions;
 using Engine.Graphics;
 
 namespace Adventure.Components
