+++ Engine/Graphics/AnimatedSprite.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Engine.Extensions;
 
 namespace Engine.Graphics
 {
