+++ Engine/Graphics/BackBuffer.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Engine.Extensions;
 
 namespace Engine.Graphics
 {
