+++ Engine/Graphics/Particles.cs
@@ -2,7 +2,6 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
-using Engine.Extensions;
 
 namespace Engine.Graphics
 {
